Work around ARM and IAR optimization issue
Add a custom profile to work around an issue with the ARM and IAR
compilers, where the example appears to hang without any error. Building
the example with `-O1` (ARM) and `-Ol` (IAR), rather than the default of
`-Os` (ARM) and `-Ohz` (IAR), works as a workaround.

To use the custom profile with mbed-cli, use the --profile option.

    mbed compile -m CY8CKIT_062_WIFI_BT -t ARMC6 --profile mbed_profile.json

Arm internal reference: SDCTRESPONSE-3550